% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%btn\%}
\alias{\%btn\%}
\title{Check whether \code{x} is in the interval given}
\usage{
x \%btn\% interval
}
\arguments{
\item{x}{A numeric vector of values to test against the interval.}

\item{interval}{An vector of two values defining an upper and lower limit.}

\item{inclusivity}{One of "[]", "()", "[)" or "(]" which are short-hand
respectively for inclusive, exclusive, left-inclusive and right-exclusive, and
left-exclusive and right-inclusive.

[)" or "(]: R:)\%22\%20or\%20\%22(}
}
\description{
\verb{\%btn\%} stands for "between" and it tests whether \code{x} is between
\code{interval[1]} and \code{interval[2]}.
}
\details{
The default behavior is to test if
\verb{x \%in\% [ interval[1], interval[2] )}, but
to get other comparisons one can call \verb{\%btn\%} without
using infix notation and use the \code{inclusivity} argument.
}
\examples{
5 \%btn\% c(0, 10)
c(0.5, 1.5) \%btn\% c(0, 1)
`\%btn\%`(c(0.5, 1), c(0, 1), inclusivity = "[]")

}
