% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%btn()\%}
\alias{\%btn()\%}
\title{Exclusive Between}
\usage{
x \%btn()\% interval
}
\arguments{
\item{x}{A numeric vector of values to test against the interval.}

\item{interval}{An vector of two values defining an upper and lower limit.}
}
\description{
Exclusive Between
}
