% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMTP_Data_Struct.R
\name{LMTP_Data_Struct}
\alias{LMTP_Data_Struct}
\title{Structured Data Object for Estimating Longitudinal Modified Treatment Policies}
\description{
The goal of having \code{LMTP_Data_Struct} is to have a standardized way to
refer to the time-varying covariates, exposure, and outcome data used in
estimating an LMTP causal parameter.
}
\examples{

df <- data.frame(
  id = 1:5,
  At1 = rnorm(n=5),
  At2 = rnorm(n=5),
  L1t1 = rnorm(n=5),
  L2t1 = rnorm(n=5),
  L1t2 = rnorm(n=5),
  L2t2 = rnorm(n=5),
  Y = rnorm(n=5)
  )

lmtp_data <- LMTP_Data_Struct$new(
  data = df,
  id_col = 'id',
  n_timesteps = 2,
  exposure_cols = c('At1', 'At2'),
  time_varying_covariate_cols = list(c('L1t1', 'L2t1'), c('L1t2', 'L2t2')),
  outcome_col = 'Y'
  )

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LMTP_Data_Struct-new}{\code{LMTP_Data_Struct$new()}}
\item \href{#method-LMTP_Data_Struct-H_cols}{\code{LMTP_Data_Struct$H_cols()}}
\item \href{#method-LMTP_Data_Struct-clone}{\code{LMTP_Data_Struct$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LMTP_Data_Struct-new"></a>}}
\if{latex}{\out{\hypertarget{method-LMTP_Data_Struct-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMTP_Data_Struct$new(
  data,
  id_col,
  n_timesteps,
  exposure_cols,
  baseline_covariate_cols = NULL,
  time_varying_covariate_cols = NULL,
  outcome_col
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LMTP_Data_Struct-H_cols"></a>}}
\if{latex}{\out{\hypertarget{method-LMTP_Data_Struct-H_cols}{}}}
\subsection{Method \code{H_cols()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMTP_Data_Struct$H_cols(t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LMTP_Data_Struct-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LMTP_Data_Struct-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMTP_Data_Struct$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
