% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmtpi_task.R
\name{LMTPI_Task}
\alias{LMTPI_Task}
\title{lmtpiTask Class Constructor}
\description{
An lmtpiTask is an object that stores the relevant setup and methods for
estimating and constructing confidence intervals for a modified treatment
policy effect.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-lmtpiTask-new}{\code{LMTPI_Task$new()}}
\item \href{#method-lmtpiTask-fit}{\code{LMTPI_Task$fit()}}
\item \href{#method-lmtpiTask-clone}{\code{LMTPI_Task$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lmtpiTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-lmtpiTask-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMTPI_Task$new(
  data = NULL,
  policy = NULL,
  config = list(),
  network = NULL,
  exposure_map = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lmtpiTask-fit"></a>}}
\if{latex}{\out{\hypertarget{method-lmtpiTask-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMTPI_Task$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lmtpiTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-lmtpiTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LMTPI_Task$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
