% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rmse}
\alias{rmse}
\title{Root-Mean Squared Error}
\usage{
rmse(x)
}
\arguments{
\item{x}{A numeric vector to take the square, then mean, and then
square-root of.}
}
\description{
Root-Mean Squared Error
}
